Fixed implementation from the first
I had an issue with the previous implementation branch wherein the times
for the output were way too long. 840ms vs 27ms. Big diff. Fixed in this
one. Not too sure how.

Can be accessed from calling python, but will need to put a few more
methods in and determine if I would like to compartmentalize it further
and introduce multiple classes for each mapspace.